+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Make DNS parameters required
+
 # 3.20.3
 
 * Adding KeyName parameter to ASG